[fix] moved .py files from lib to main folder
