+++ triangle.py
